Add support for void cast
We model this as an expression, just like all the other casts, even
though PHP treats (void) specially and makes it a statement. However,
it has a peculiar carve-out where (void) *is* allowed in for init
and loop expression lists (which only accept expressions, not
statements). As such, we need to model it as an expression.